Add initial C++ project sources
It's a good idea to not add too much C++ right now until we have all the
other elements of the project worked out, and are happy with it. Large
codebases are harder to refactor in major ways than smaller codebases...

Headers are added under an include/ directory, like a standard C++
project. We place the header under an additional directory for the
project to namespace the header file and prevent name collisions.
C++ source files are added under the src/, like a standard C++ project.

We have not added a build system yet, but we can still compile the
library manually.
Assuming a Unix-like system with GCC or Clang:
> c++ -Iinclude/ -fPIC -c src/hello.cpp -o hello.o
> c++ --shared hello.o -o hello.so