Run tests with Travis CI
Sign up for Travis CI at travis-ci.org. Then we enable the "hello"
project in the Travis CI GUI. Finally, we add a .travis.yml file which
instructs Travis on how to run the tests, Nothing else is needed to get
Travis to run on all local branch pushes or PRs opened in forks. We add
a line to the Travis configuration to ensure only the master local
branch is tested since we choose the PR-flow.