Use tox for project automation
Up to this point we have been manually running test and documentation
build processes. Sometimes these processes can be quite cumbersome. Tox
was created to automate testing processes, but can also be used to
automate documentation building as well.

tox is configured using the tox.ini file. In it we declare multiple
'testenv's that configure an isolated environment to run commands in.
We have set up a default testenv for running our Python tests. More
testenvs will come.

To run the tests
> tox -e test