Add pybind11 binding and CMake build process
The pybind11 binding for the C++ library is added under src/pyhello.
The 'py' prefix is common for Python wrappers around existing libraries.
The sources are placed in the similarly named directory under src to
segregate them, but also because this directory will be the eventual
location of the Python package.

See the CMakeLists.txt for details on what was changed with the build
system.

Building the pybind11 module is the same as the rest of the library:
> cmake -S . -B build
> cmake --build build

The _pyhello extension module will be located in the build/ directory.
You can test it manually.
> python -i
>> import build._pyhello as hello
>> hello.hello()
>> "Hello, World!"