Add Python package and scikit-build support
We add a toplevel __init__.py for the hello package.

To support Python packaging we add a pyproject.toml for PEP518 support
and split setuptools configuration between a setup.py and setup.cfg.
Ideally, we could put everything in a pyproject.toml, but scikit-build
doesn't support that yet.

See the CMakeLists.txt for details on the change to the build process.
This included moving the pybind wrapper build into a conditional section
so it is only build when CMake is invoked by scikit-build.

You can install the Python package, which will also install the C++
headers and CMake project files into the Python package prefix.
> pip install .