Setup Sphinx documention
Sphinx is a more powerful documentation tool than Doxygen, and with the
help of the Breathe extension we can use the existing doxygen in-source
documentation in Sphinx. In addition, we can document the Python
wrapper library as well.

To generate the sphinx documentation first generate the doygen
documentation
> cd docs
> doxygen Doxyfile

Then you can build the sphinx documentation
> sphinx -b . out