Add Python module documentation to Sphinx
Previously we only documented the C++ interface, but now we use the
autodoc extension to sphinx to generate documentation for the Python
modules.

Things get a bit tricky here because we are trying to get a python
extension module to add like a package __init__.py. Hence the __all__
and __doc__ import in __init__.py. This probably won't be appropriate
for your use case.

Documentation is still built the same way
> tox -e docs

This time an additional section is added to the documentation that
documents all the members of the hello module automagically.