Add support for readthedocs
Added a readthedocs configuration file (.readthedocs.yml) that set the
necessary settings for readthedocs to run sphinx in the docs directory.
Readthedocs does not support arbitrary execution steps so we can't run
doxygen first in a normal way, so we do it in a less-than-normal way by
running it as a subprocess in Sphinx's conf.py which means it will run
before the Sphinx documentation is generated. See
https://breathe.readthedocs.io/en/latest/readthedocs.html
for more details.

We also remove the doxygen build step from tox andadd the documentation
badge to the README.

Sign up on readthedocs.org. Once complete, register your project with
ReadTheDocs. It will then pick up a .readthedocs.yml file to use to
direct the building of your documentation. Check the settings to
configure when docs are built.